refactor(v0.6): split core/reasoning/reflect.py 29.2KB → 5-file package
Per CLAUDE.md rule #5 ("no file in `core/` exceeds 20 KB. If your
change pushes a file over, split first."). v0.6 entry skeleton §5
NEW solo-doable items + v0.5 close handover doc-consistency audit
(#887) flagged 5 oversize legacy core/ modules. This PR closes the
largest one — `core/reasoning/reflect.py` at 29.2 KB.

## The split

Pre-split:
  core/reasoning/reflect.py  (29.2 KB / 697 lines)

Post-split:
  core/reasoning/reflect/
    __init__.py          (4.4 KB) — re-export façade + singleton
    prompts.py           (8.4 KB) — CRITIQUE/REVISE_PROMPT_* + caps
    meta_narration.py    (4.4 KB) — pattern detector + stripper
    issue_extractor.py   (2.7 KB) — critique → tag map
    loop.py             (14.6 KB) — ReflectionLoop class + helpers

All 5 sub-files under the 20 KB cap. Logical groupings preserved
the cohesion-coupling boundaries that already existed in the source
(prompts grouped, meta-narration helpers grouped, etc.).

## Behaviour preservation

The split is a no-op for callers. Every import path the pre-split
file supported keeps working through the `__init__.py` façade:

  Public:
    ReflectionLoop, get_reflection_loop, DEFAULT_BACKEND_ID

  Private (test-only):
    _clear_singleton_for_tests, _enabled, _no_issues,
    _extract_issue_flag, _looks_like_meta_narration,
    _strip_meta_narration,
    _META_NARRATIVE_PATTERNS, _ISSUE_TYPE_PATTERNS

  Constants (research scripts + tests):
    CRITIQUE_PROMPT_EN/KO, REVISE_PROMPT_EN/KO,
    REVISE_PROMPT_V2_EN/KO,
    DEFAULT_CRITIQUE_TIMEOUT_S, DEFAULT_REVISE_TIMEOUT_S,
    DEFAULT_CRITIQUE_MAX_TOKENS (=4096),
    DEFAULT_REVISE_MAX_TOKENS (=1024),
    MAX_REVISE_RATIO (=2.5)

The singleton (`_SINGLETON` + `_SINGLETON_LOCK` + `get_reflection_loop`
+ `_clear_singleton_for_tests`) lives in `__init__.py` rather than
in `loop.py` so the import path `from core.reasoning.reflect import
get_reflection_loop` keeps working without forcing callers to update
to `from core.reasoning.reflect.loop import get_reflection_loop`.

## Module docstring promotion

The legacy file's module-level docstring (the canonical
ARCHITECTURE.md §5.7.1 reference + opt-in posture + JAMES_ENABLE_REFLECT
+ trace row contract + wiring note) moved to the package's
`__init__.py` docstring. A new "v0.6 package split" subsection
documents the file-level decomposition so a future contributor sees
the boundary at a glance.

## feature_flags.py path reference

`core/feature_flags.py` line 73 had a string reference
`"module": "core/reasoning/reflect.py"`. Updated to
`"core/reasoning/reflect/"` so operator-facing feature-flag
documentation (which reads this field) points at the package
directory rather than a non-existent file.

## Tests — tests/test_v06_reflect_module_size.py (NEW, 8 tests)

  * `ModuleSizeCapTests` (4):
      - legacy single-file removed (so it can't regrow)
      - package directory exists
      - 5 canonical sub-files present by exact name
      - every sub-file under 20 KB (CLAUDE.md rule #5 lock)
  * `PublicImportSurfaceTests` (3):
      - canonical public imports (ReflectionLoop +
        get_reflection_loop + DEFAULT_BACKEND_ID)
      - canonical private imports (8 symbols that tests depend on)
      - canonical prompt constants — smoke-checks template content
        ("Contradiction" / "모순" / "{issue_type}") + cap defaults
        (4096 / 1024 / 2.5)
  * `SingletonContractTests` (1):
      - lazy + cached construction; _clear_singleton_for_tests
        reset works

8/8 pass.

Regression sweep on the existing reflect-related test suites
(test_reflection_loop + test_reflect_meta_narration_strip +
test_reflect_d1_wiring + test_sector_flag_s6_cognitive): 31/31 pass.
Combined with the new lock-tests: 39/39 pass.

## Verification

- pytest tests/test_v06_reflect_module_size.py: 8 passed
- pytest reflect-related regression suite: 31 passed
- core/retrieval / core/graph traversal / core/reasoning lines
  changed: behaviour preserved byte-identical — only file location
  changed
- All 22 import paths (public + private + constants) verified to
  resolve to the same Python objects as the pre-split file
- pipeline_synth.py importing get_reflection_loop verified
- feature_flags.py reference path updated

## What this PR does NOT do

- Does NOT change behaviour. This is a pure refactor — every
  function does what it did before. The split test suite (39 tests)
  validates this
- Does NOT touch the 4 remaining oversize files (`gemma_client.py`
  23.7 KB / `replay_graph.py` 23.0 KB / `_ingestion.py` 22.5 KB /
  `_frontmatter.py` 21.4 KB). Each will land in its own PR for
  reviewability — bundling all 5 into one PR makes the diff
  unreadable
- Does NOT introduce any vertical content — BLOCKED per CLAUDE.md
  rule #1
- Does NOT modify any test that depends on the reflect module —
  tests pass as-is with the old import paths

Quality delta: exempt (label: refactor — file reorganisation +
size-cap lock tests; no `core/retrieval`, `core/graph` traversal,
or `core/reasoning` behaviour changes; all 22 import paths
preserved)
